ai/mxdataset_devkit/pyproject.toml
[build-system]
requires = ["setuptools>=43.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mxdataset_devkit"
# version = "0.0.0"
dependencies = [
  "pyyaml",
  "requests",
  "boto3",
  "lakefs_client @ git+http://192.168.70.202:30453/ai-model/lakefs_client.git@master",
  "pycocotools>=2.0.6",
]
requires-python = ">=3.8"
authors = [
  {name = "zwx", email = "zwxdlut@163.com"},
]
maintainers = [
  {name = "zwx", email = "zwxdlut@163.com"}
]
description = "A mxdataset devkit."
# readme = "README.md"
license = {file = "LICENSE"}
keywords = ["mxdataset", "devkit", "s3", "lakefs"]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python",
  "Development Status :: 3 - Alpha",
  "Environment :: Console",
  "Framework :: Flask",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: GPLv3 License",
  "Operating System :: OS Independent",
  "Programming Language :: Python :: 3",
  "Topic :: Software Development :: Libraries :: Python Modules"
]
dynamic = ["version", "readme"]

# [project.optional-dependencies]
# gui = ["PyQt5"]
# cli = [
#   "rich",
#   "click",
# ]
# pdf = ["ReportLab>=1.2", "RXP"]
# rest = ["docutils>=0.3", "pack ==1.1, ==1.3"]

# [project.urls]
# Homepage = "https://example.com"
# Documentation = "https://readthedocs.org"
# Repository = "https://github.com/me/spam.git"
# "Bug Tracker" = "https://github.com/me/spam/issues"
# Changelog = "https://github.com/me/spam/blob/master/CHANGELOG.md"

[project.scripts]
# my-script = "my_package.module:function"
migrate = "mxdataset.migration.migration:migrate"

# [project.gui-scripts]
# spam-gui = "spam:main_gui"

# [project.entry-points."spam.magical"]
# tomatoes = "spam:main_tomatoes"

# [tool.setuptools]
# packages = ["mypkg", "mypkg.subpkg1", "mypkg.subpkg2"]
# package-dir = {"" = "src"}
#     # directory containing all the packages (e.g.  src/mypkg1, src/mypkg2)

[tool.setuptools.packages]
find = {}  # Scanning implicit namespaces is active by default
# # OR
# find = {namespaces = false}  # Disable implicit namespaces

# [tool.setuptools.package-dir]
# mypkg = "lib"
# # mypkg.module corresponds to lib/module.py
# "mypkg.subpkg1" = "lib1"
# # mypkg.subpkg1.module1 corresponds to lib1/module1.py
# "mypkg.subpkg2" = "lib2"
# # mypkg.subpkg2.module2 corresponds to lib2/module2.py

# [tool.setuptools.packages.find]
# where = ["src"]  # list of folders that contain the packages (["."] by default)
# include = ["my_package*"]  # package names should match these glob patterns (["*"] by default)
# exclude = ["my_package.tests*"]  # exclude packages matching these glob patterns (empty by default)
# namespaces = false  # to disable scanning PEP 420 namespaces (true by default)

[tool.setuptools.dynamic]
version = {attr = "mxdataset.__version__"}  # any module attribute compatible with ast.literal_eval
readme = {file = ["README.md"]}